pyproject.toml
[tool.poetry]
name = "DRF_LMS_project"
version = "0.1.0"
description = ""
authors = ["Sergey Spisak <serega232003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
ipython = "^8.28.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
